+++ tests/TestServer/index.php
@@ -1,0 +1,11 @@
+<?php
+
+use Busarm\PhpMini\Server;
+use Busarm\PhpMini\Test\ServerTest;
+
+require '../bootstrap.php';
+
+$server = (new Server())
+    ->addRoutePath('v1', __DIR__ . '/../TestApp')
+    ->addDomainPath('localhost:' . ServerTest::HTTP_TEST_PORT, __DIR__ . '/../TestApp');
+$server->run();